Check path before removeing packages directory
- fix the issue when running install_packages_to_data_dir.ps1
  on Windows.

+++ waspc/tools/install_packages_to_data_dir.ps1
@@ -29,7 +29,9 @@ foreach ($package in $packageDirs) {
 
 Push-Location "$dir/.."
 try {
-    Remove-Item -Path "./data/packages" -Recurse -Force
+    if (Test-Path "./data/packages") {
+        Remove-Item -Path "./data/packages" -Recurse -Force
+    }
     Copy-Item -Path "./packages" -Destination "./data/packages" -Recurse
 }
 finally {
